Extract legacy aws models
